extract logic to dedicated WannabeOnRequest handler
